requirements.txt
fastapi==0.115.8
uvicorn==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
bcrypt
openai==1.55.3
pypdf==4.2.0
numpy==2.1.3
python-dotenv==1.0.1
itsdangerous==2.2.0
pytest==8.3.3
